@@ -10,13 +10,14 @@
   },
   "dependencies": {
     "@dipakparmar/docusaurus-plugin-umami": "^2.0.6",
-    "@docusaurus/core": "2.4.1",
-    "@docusaurus/preset-classic": "2.4.1",
+    "@docusaurus/core": "3.6.3",
+    "@docusaurus/preset-classic": "3.6.3",
+    "@getcanary/docusaurus-theme-search-pagefind": "^1.0.2",
+    "@getcanary/web": "^1.0.12",
     "classnames": "^2.2.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-lite-youtube-embed": "^2.0.3",
-    "url-search-params-polyfill": "^8.1.0"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-lite-youtube-embed": "^2.0.3"
   },
   "devDependencies": {
     "netlify-plugin-cache": "^1.0.3"
