@@ -1,4 +1,6 @@
-module.exports = {
+import type { SidebarsConfig } from '@docusaurus/plugin-content-docs'
+
+const sidebars: SidebarsConfig = {
   docs: [
     {
       type: 'category',
@@ -42,3 +44,5 @@ module.exports = {
     },
   ],
 }
+
+export default sidebars
